src/py/requirements.txt
mysql-connector-python~=8.0.32
boto3
botocore